【WA】[Kyuride Kagamiz Programming Contest (Easy)]D - 虫歯(Cavity)

+++ k2pc-easy/k2pc001_e4.cpp
@@ -37,7 +37,7 @@ bool contain(set<ll> &s, int a) { return s.find(a) != s.end(); }
 
 typedef priority_queue<ll, vector<ll>, greater<ll> > PQ_ASK;
 
-ll to_id(int p, int q) {
+ll to_id(ll p, ll q) {
     // 0,1 -> 0
     // 1,1 -> 1
     // 1,2 -> 2
